Refactor remove segment
